+++ composeApp/src/commonMain/kotlin/com/gideon/dogify/App.kt
@@ -1,27 +1,14 @@
 package com.gideon.dogify
 
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.safeContentPadding
-import androidx.compose.material3.Button
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import com.gideon.dogify.model.Breed
-import com.gideon.dogify.usecase.FetchBreedsUseCase
-import com.gideon.dogify.usecase.GetBreedsUseCase
-import com.gideon.dogify.usecase.ToggleFavouriteStateUseCase
-import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.ui.tooling.preview.Preview
-
-import dogify.composeapp.generated.resources.Res
-import dogify.composeapp.generated.resources.compose_multiplatform
 
 @Composable
 @Preview