+++ composeApp/src/commonMain/kotlin/com/gideon/dogify/api/model/Breed.kt
@@ -1,3 +1,3 @@
-package com.gideon.dogify.model
+package com.gideon.dogify.api.model
 
 data class Breed(val name: String, val imageUrl: String, val isFavourite: Boolean = false)