+++ composeApp/src/commonMain/kotlin/com/gideon/dogify/database/DriverFactory.kt
@@ -1,0 +1,6 @@
+package com.gideon.dogify.database
+
+import app.cash.sqldelight.db.SqlDriver
+import org.koin.core.scope.Scope
+
+internal expect fun Scope.createDriver(databaseName: String): SqlDriver