+++ composeApp/src/commonMain/kotlin/com/gideon/dogify/di/KoinModule.kt
@@ -1,11 +1,33 @@
 package com.gideon.dogify.di
 
+import com.gideon.dogify.api.BreedsApi
+import com.gideon.dogify.database.createDriver
+import com.gideon.dogify.db.DogifyDatabase
+import com.gideon.dogify.repository.BreedRemoteSource
+import com.gideon.dogify.repository.BreedsLocalSource
+import com.gideon.dogify.repository.BreedsRepository
 import com.gideon.dogify.usecase.FetchBreedsUseCase
 import com.gideon.dogify.usecase.GetBreedsUseCase
 import com.gideon.dogify.usecase.ToggleFavouriteStateUseCase
+import com.gideon.dogify.util.getDispatcherProvider
 import org.koin.core.context.startKoin
 import org.koin.dsl.KoinAppDeclaration
 import org.koin.dsl.module
+
+private val utilityModule = module {
+    factory { getDispatcherProvider() }
+    single { DogifyDatabase(createDriver("dogify.db")) }
+}
+
+private val apiModule = module {
+    factory { BreedsApi() }
+}
+
+private val repositoryModule = module {
+    single { BreedsRepository(get(), get()) }
+    factory { BreedRemoteSource(get(), get()) }
+    factory { BreedsLocalSource(get(), get()) }
+}
 
 private val useCaseModule = module {
     factory { GetBreedsUseCase() }
@@ -13,7 +35,7 @@ private val useCaseModule = module {
     factory { ToggleFavouriteStateUseCase() }
 }
 
-private val sharedModules = listOf(useCaseModule)
+private val sharedModules = listOf(useCaseModule, repositoryModule, apiModule, utilityModule)
 
 fun initKoin(appDeclaration: KoinAppDeclaration = {}) = startKoin {
     appDeclaration()