+++ composeApp/src/commonMain/kotlin/com/gideon/dogify/usecase/FetchBreedsUseCase.kt
@@ -1,7 +1,12 @@
 package com.gideon.dogify.usecase
 
-import com.gideon.dogify.model.Breed
+import com.gideon.dogify.api.model.Breed
+import com.gideon.dogify.repository.BreedsRepository
+import org.koin.core.component.KoinComponent
+import org.koin.core.component.inject
 
-class FetchBreedsUseCase {
-    suspend fun invoke(): List<Breed> = listOf(Breed("Test fetch", ""))
+class FetchBreedsUseCase : KoinComponent {
+    private val breedsRepository: BreedsRepository by inject()
+
+    suspend fun invoke(): List<Breed> = breedsRepository.fetch()
 }