+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-02/12/2025 - ch05/01
+05/12/2025
+- add kotlin coroutines
+- add db connection and fix bug on serialization plugin 
+
+03/12/2025
+- add sqlDelight dependency
+
+02/12/2025
 - init project
 - add dependencies
 - add license 
+- fetch to dogceo api